Refector: staticRate type to Long

+++ src/main/kotlin/com/hwalaon/wezxro_server/domain/account/controller/response/AccountDetailResponse.kt
@@ -9,7 +9,7 @@ data class AccountDetailResponse (
     val email: String,
     val money: Double,
     val status: BasicStatus,
-    val staticRate: Long,
+    val staticRate: Float,
 ) {
     companion object {
         fun fromDomain(domain: Account) =
